# swagger/swagger.yml
swagger: "2.0"
info:
  description: ""
  version: "0.0.1"
  title: "prodx"
  contact:
    email: "alexandre@ineosystems.com.br"
host: "127.0.0.1:3000"
basePath: "/v1"
...
schemes:
- "http"
securityDefinitions:
  APIKeyHeader:
    type: apiKey
    in: header
    name: x-api-token
paths:
...
  /cell:
    post:
      tags:
      - "cell"
      summary: "Add a new cell"
      description: ""
      operationId: "addCell"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      security:
        - APIKeyHeader: []
      parameters:
      - in: "body"
        name: "body"
        description: "Cell object that needs to be added"
        required: true
        schema:
          $ref: "#/definitions/Cell"
      responses:
        201:
          description: "Created"
          schema:
            type: integer
        405:
          description: "Invalid input"
          schema:
            $ref: "#definitions/ApiResponse"
        409:
          description: "Already exists"
          schema:
            $ref: "#definitions/ApiResponse"
        500:
          description: "Internal error"
          schema:
            $ref: "#definitions/ApiResponse"
  /cells:
    get:
      tags:
      - "cell"
      summary: "Finds Cells by customer"
      operationId: "findCellByCustomer"
      produces:
      - "application/json"
      security:
        - APIKeyHeader: []
      responses:
        200:
          description: "successful operation"
          schema:
            type: array
            items:
              $ref: "#/definitions/Cell"
        400:
          description: "Invalid cell name"
        404:
          description: "cells not found"
        500:
          description: "Internal error"
          schema:
            $ref: "#definitions/ApiResponse"
  /cell/{cell_id}:
    get:
      tags:
      - "cell"
      summary: "Find cell by ID"
      description: "Returns a single cell"
      operationId: "getCellById"
      produces:
      - "application/json"
      security:
        - APIKeyHeader: []
      parameters:
      - name: "cell_id"
        in: "path"
        description: "ID of cell to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Cell"
        404:
          description: "Cell not found"
        405:
          description: "Invalid input"
          schema:
            $ref: "#definitions/ApiResponse"
        500:
          description: "Internal error"
          schema:
            $ref: "#definitions/ApiResponse"
    post:
      tags:
      - "cell"
      summary: "Updates a cell with form data"
      description: ""
      operationId: "updateCellWithForm"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      security:
        - APIKeyHeader: []
      parameters:
      - name: "cell_id"
        in: "path"
        description: "ID of cell that needs to be updated"
        required: true
        type: "integer"
        format: "int64"
      - name: "public_key"
        in: "formData"
        description: "Updated name of the cell"
        required: false
        type: "string"
      responses:
        405:
          description: "Invalid input"
    delete:
      tags:
      - "cell"
      summary: "Deletes a cell"
      description: ""
      operationId: "deleteCell"
      produces:
      - "application/json"
      security:
        - APIKeyHeader: []
      parameters:
      - name: "api_key"
        in: "header"
        required: false
        type: "string"
      - name: "cell_id"
        in: "path"
        description: "Cell id to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        400:
          description: "Invalid ID supplied"
        404:
          description: "Cell not found"
...
definitions:
...
  Cell:
    type: object
    required:
      - name
    properties:
      id:
        type: integer
      name:
        type: string
...